backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-chess>=1.10.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic-ai>=0.0.24
supabase>=2.3.0
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
